Update for latest Spree
There are commits on the `3.0-stable` branch that moves this addon to
the at 3.0 RC5, but we want the ability to be on 3.x stables so not
even using that.

The goal of this commit is to not change any functionality but just
update enough that I can write running specs. This means:

* Updated gemspec to use any 3.0 version of spree.
* Replace all generate files with new versions. This includes:
  * Rakefile
  * bin/rails
  * spec/spec_helper.rb
  * spree_api_auth.gemspec
  * controller_setup.rb - removed
* I removed the install generator as the dummy app (and presumably
  spree stores) would use it and its non-working. Since it's mostly
  about integrating assets and copying migrations (neither of which we
  have) I just removed it.
* I bumped the min ruby version to 2.1.0 since that is the min version
  of Spree 3.
* I remove the debugger dependency as both gems that are not out of
  date and replaced with `byebug` (which is compatible with Ruby > 2.1)

The final bit of trickery I did was to apply the decorator to either
Spree::Api::V1::UsersController OR Spree::Api::UsersController
(whichever exists). This allows it to be compatible with Spree 3.1
(which uses the former) and Spree 3.0 (which uses the later).

I might later just get rid of the decorator usage and have a custom
controller but that's out of scope for this commit. Need to have working
tests first.