Implement automated tests for future development
Although my goal was not to break compatibility yet the `v1` namespace
of Spree 3.1 prompted me to. My tests revealed that my last commit would
get us good on Spree 3.0 but not Spree 3.1.

Spree 3.1, in an attempt to redirect folks to the `v1` namespace, has
implemented a global redirect. This redirect is misguided as it really
only makes sense for GET requests and assumes clients will follow
redirects. But regardless it is what exists.

This means it will try to redirect before the non-`v1` route this plugin
adds is evaluated. In an attempt to fix this I changed everything in
this plugin to be `v1` namespaced.

Turns out I still had the same problem. The redirect is happening before
the add-on route is evaluated. To really solve the problem I had to
update `add_routes` to `routes.draw` in the routes file. This may lead
to reloading issues later but it seems what is necessary for now.

This means I didn't strictly have to break compat, but since I am
planning to do that anyway might as well be forward looking.