Remove unnecessary templates
It appears the templates were copied from the stock implementation
rather than just rendering the `show`. This means it won't take
advantage of advancements (such as the caching added) or customizations
(such as custom attributes added).

The only reason for this customization is to add the API key to the
response. We can do that via the API helper hook so doing that instead
and getting rid of all the templates except the user_exists one.

Also made the user_exists one relative so it will work in both Spree 3.0
and Spree 3.1.