Bind add-on to devise
This add-on does assume devise is our authentication mechanism. Rather
than using Spree::LegacyUser (the default in testing a gem) we want to
make the connection to Devise explicit.

This will come in useful as we adjust the code to correctly validate
the password through devise.

As part of this we had to bump back down to Rails 4. The latest stable
version of spree_auth_devise is not compatible with a version of devise
which is compatible with Rails 5.

Due to the bump down in to rails 4 we had to update the tests to not
explicitly provide the params.