Use correct authentication mechanism
Previously Spree::User#valid_password? was used to verify the password.
While this works IF we are only using the database authentication
strategy (the default) it doesn't work if we have a custom
authentication strategy registered with warden.

This means under this configuration, even a valid password would not
authenticate. To make this work I had to break the API by having the
params be `spree_user` instead of `user` which is what warden has been
configured to expect by spree_auth_devise.