Fix Spree 3.0 compat for controller namespace
I thought my constant handing would have worked for 3.0 but turns out
that Spree::Api::V1::UsersController is already defined somehow. It
actually returns a reference to Spree::Api::UsersController.

I did a bit of research but couldn't find where this alias is actually
defined. It appears in 2012 Spree had a `V1` namespace. Then it got
unfashionable and was replaced with version_cake. Then it got
fashionable again in 2015 and the reversed course.

I'm guessing somewhere in there an alias compatibility was added but
I'm not completely sure.

Anyway, to satisfy the autoloading we take the found constant and define
it as a real constant. At that point we have something that work on
both Spree 3.0 and Spree >= 3.1.