Replace generic 401 unauth message with devise specific one
The generic message can be a little vague. "You are not authorized to
perform that action" doesn't make it clear that the credentials that
were given were not valid.

To fix I am having it read from the devise locale messgae. This will
ensure the message given via the API is consistent with the website.